% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/knitron.R
\name{eng_ipython}
\alias{eng_ipython}
\title{An IPython engine that can be registered with knitr}
\usage{
eng_ipython(options)
}
\arguments{
\item{options}{an knitr option list}
}
\value{
output for knitr
}
\description{
An IPython engine that can be registered with knitr
}

