% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/knitron.R
\name{knitron.execute_chunk}
\alias{knitron.execute_chunk}
\title{Execute a code chunk from a knitr option list}
\usage{
knitron.execute_chunk(options, profile = "knitr")
}
\arguments{
\item{options}{a knitr option list}

\item{profile}{the name of the profile}
}
\value{
a data frame of messages from the Python wrapper
}
\description{
Execute a code chunk from a knitr option list
}

