% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/knitron.R
\name{knitron.execute_code}
\alias{knitron.execute_code}
\title{Execute a single Python command}
\usage{
knitron.execute_code(code, profile = "knitr")
}
\arguments{
\item{code}{the command to execute}

\item{profile}{the name of the profile}
}
\value{
the command's stdout and stderr
}
\description{
Execute a single Python command
}

