% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/knitron.R
\name{knitron.is_running}
\alias{knitron.is_running}
\title{Returns true if the cluster is running}
\usage{
knitron.is_running(profile = "knitr")
}
\arguments{
\item{profile}{the name of the profile}
}
\value{
\code{TRUE} if cluster is running
}
\description{
Returns true if the cluster is running
}

