% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{knitron}
\alias{knitron}
\alias{knitron-package}
\title{IPython and Matplotlib integration for knitr}
\description{
After loading this package, you can make use of the \code{engine = 'ipython'}
chunk option in \pkg{knitr}.
}
\author{
Fabian Hirschmann <\url{http://0x0b.de}>
}

