% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/knitron.R
\name{knitron.start}
\alias{knitron.start}
\title{Starts an IPython cluster with one engine}
\usage{
knitron.start(profile = "knitr", wait = TRUE, quiet = TRUE)
}
\arguments{
\item{profile}{the name of the profile}

\item{wait}{wait for the engine to start up}

\item{quiet}{be quiet about IPython's startup messages}
}
\description{
Starts an IPython cluster with one engine
}

