% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/knitron.R
\name{knitron.stop}
\alias{knitron.stop}
\title{Stops an IPython cluster}
\usage{
knitron.stop(profile = "knitr", quiet = TRUE)
}
\arguments{
\item{profile}{the name of the profile}

\item{quiet}{be quiet about IPython's shutdown messages}
}
\description{
Stops an IPython cluster
}

